+++ Sources/VComponents/Models/ModalComponentSizes.swift
@@ -42,14 +42,10 @@ public struct ModalComponentSizes<ModalSize> {
         }
     }
 
-    @available(macOS, unavailable)
-    @available(tvOS, unavailable)
-    @available(watchOS, unavailable)
-    @available(visionOS, unavailable)
     func current(
-        _interfaceOrientation interfaceOrientation: _InterfaceOrientation
+        orientation: PlatformInterfaceOrientation
     ) -> ModalSize {
-        switch interfaceOrientation {
+        switch orientation {
         case .portrait: portrait
         case .landscape: landscape
         }