+++ CHANGELOG.md
@@ -19,6 +19,7 @@ VDisclosureGroup
 
 VModal
 
+- Modal now supports `macOS`, `tvOS`, and `visionOS`
 - `roundedCorners` and `cornerRadius` are replaced with `cornerRadii`
 
 VSideBar
@@ -48,9 +49,9 @@ VAlert
 
 VToast
 
+- Toast can now be dismissed with pull-down
 - API is overhauled
 - Modal no longer requires `ColorScheme` as parameter when it's overridden in the root
-- Toast can now be dismissed from pull-down
 
 ### [6.0.7(67)](https://github.com/VakhoKontridze/VComponents/releases/tag/6.0.7) — *2024 07 09*